Move Term struct to its own file
